Code refactoring №3
