+++ CraftingGillionaire/Models/Marketshare/MarketshareFilterItem.cs
@@ -4,18 +4,18 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace CraftingGillionaire.Models
+namespace CraftingGillionaire.Models.Marketshare
 {
     public class MarketshareFilterItem
     {
-        public MarketshareFilterItem(string filterName, int filterID) 
-        { 
+        public MarketshareFilterItem(string filterName, int filterID)
+        {
             this.Name = filterName;
             this.ID = filterID;
         }
 
-        public string Name { get; set; }
+        public string Name { get; }
 
-        public int ID { get; set; }
+        public int ID { get; }
     }
 }
