+++ CraftingGillionaire/Models/SalesHistory/SaleDisplayItem.cs
